requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
black==21.9b0
cfgv==3.3.1
click==8.0.3
cron-descriptor==1.4.5
croniter==1.0.15
distlib==0.3.3
fastapi==0.68.2
filelock==3.3.0
greenlet==1.1.2
h11==0.12.0
identify==2.3.0
Jinja2==3.0.2
MarkupSafe==2.0.1
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
platformdirs==2.4.0
pre-commit==2.15.0
psutil==5.8.0
pydantic==1.8.2
python-crontab==2.5.1
python-dateutil==2.8.2
PyYAML==5.4.1
regex==2021.10.8
six==1.16.0
SQLAlchemy==1.4.25
starlette==0.14.2
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
uvicorn==0.15.0
virtualenv==20.8.1
